+++ 04-Makefiles/codes-2024-02-23/Makefile
@@ -1,13 +1,40 @@
+#Bandera que se lo colva al compildor
+CXXFLAGS=-o2 -std=c++17 
+
 all: sumupdown.pdf
 
-sumupdown.x: sumupdown.cpp
-	g++ sumupdown.cpp -o sumupdown.x
+# sumupdown.x: sumupdown.cpp
+# 	g++  sumupdown.cpp -o sumupdown.x
 
-# %.x: %.cpp
-#  	g++ $< -o $@
+# sumupdown.x: sumupdown.cpp
+# 	g++ $(CXXFLAGS) sumupdown.cpp -o sumupdown.x
+
+#Establece la dependencia entre los archivos con extension
+#.cpp y lso archivos ejecutables con extension .x
+
+#La notacion % es un comodin que coincide con cualquier
+#cadena de caracteres
+
+# $^: Es una variable automática que representa todas
+#las dependencias. En este caso, representa el archivo
+#fuente ('%.cpp')
+
+# -o $@: Especifica el nombre del archivo de salida.
+# $@ es una variable automatica que representa el objeto
+# de la regla, en este caso, el archivo ejecutable 
+#(%x)
+
+%.x: %.cpp
+	g++ $(CXXFLAGS) $^ -o $@
 
 data.txt: sumupdown.x
 	./sumupdown.x > data.txt
 
 sumupdown.pdf: data.txt script.gp
 	gnuplot script.gp
+
+
+.PHONY: clean
+
+clean:
+	rm -f *.x *.txt *.pdf
