+++ 04-Makefiles/codes-2024-02-23/functions.h
@@ -1,0 +1,4 @@
+#pragma once
+float sumup(int nterms);
+float sumdown(int nterms);
+float sumdown(int nterms);
