+++ 04-Makefiles/codes-2024-02-23/mainsum.cpp
@@ -1,8 +1,9 @@
 #include <iostream>
 #include <cmath>
+#include "functions.h"
 
-float sumup(int nterms);
-float sumdown(int nterms);
+
+
 
 int main(int argc, char **argv)
 {
@@ -17,19 +18,3 @@ int main(int argc, char **argv)
   return 0;
 }
 
-float sumup(int nterms)
-{
-  float result = 0;
-  for(int n = 1; n <= nterms; n++) {
-    result += 1.0/n;
-  }
-  return result;
-}
-float sumdown(int nterms)
-{
-  float result = 0;
-  for(int n = nterms; n >= 1; n--) {
-    result += 1.0/n;
-  }
-  return result;
-}
